feat: create specific method for parsing csv vals
