+++ test/commands/scan/eol.test.ts
@@ -5,11 +5,11 @@ import { afterEach, beforeEach, describe, it } from 'node:test';
 import { runCommand } from '@oclif/test';
 import * as sinon from 'sinon';
 
-import { default as EolScan } from '../../../src/commands/scan/eol.ts';
 import { default as SbomScan } from '../../../src/commands/scan/sbom.ts';
-import { type Sbom, cdxgen, extractPurls, prepareRows } from '../../../src/service/eol/eol.svc.ts';
+import { type Sbom, cdxgen, prepareRows } from '../../../src/service/eol/eol.svc.ts';
 import type { CdxCreator } from '../../../src/service/eol/eol.types.ts';
 import { type ScanResponseReport, type ScanResult, buildScanResult } from '../../../src/service/nes/modules/sbom.ts';
+import { extractPurls } from '../../../src/service/purls.svc.ts';
 import { FetchMock } from '../../utils/mocks/fetch.mock.ts';
 import { InquirerMock } from '../../utils/mocks/ui.mock.ts';
